THE BIG KAHUNA: Filter map is finished
Filter now working completey.

1) Asset Category can now be selected by a user, while Aseet Type is
grayed out

2) Once Aseet Category is selected, Asset Type can be elected as well

3) Asset Filter Map's clear button now works. If a user presses it, it
will set the Filter Map back to default

4) Asset Filter Map's apply button now works. If a user presses it, they
will be prompted with a messege. If "OK" is selected. The user will be
taken back to the map view, and will be displayed assets that fulfil the
filter. If "CANCEL" is pressed nothing happens.

5) Asset Map's Filter button, changes to Clear Filter so if a user
decides to go to list and come back, the filter will still be active.

6) Layouts have been added to Filter Map, Asset Category and Asset Type