The final update from me :( (unless..)
1) If a user turns of location in settings, a prompt will ask them to
re-enable them. A user will be taken to the settings menu if they wish
to do so.

2) Each location setting now uses the priority that matches there power
consumption (i.e GPS+(Netowrk or Wifi) = Priority_High_Accuracy

3) The entire app now saves it state upon rotation. This was done in the
.xml manifest, which means we don't need to code the indiviually in each
method.

4) Added all user hints found in Main Activity to the strings.xml, also
fixed CustomAssetFilterAdapter so that  if the language changes the if
statements will still work

5) Added Block comments on the very top that explains what each class
does. CustomAssetFilterAdapter has in-line comments left for easier
viewing